Split component validation
