+++ CMakeLists.txt
@@ -214,6 +214,7 @@ target_sources(
         src/game_data_validation_controller_components.c
         src/game_data_validation_factions.c
         src/game_data_validation_brush_worlds.c
+        src/game_data_validation_components.c
         src/game_data_validation_editor_metadata.c
         src/game_data_validation_grid.c
         src/game_data_validation_haptics.c
